feat ResultSet结果映射

+++ src/main/java/com/zhengqing/demo/mapper/UserMapper.java
@@ -16,9 +16,9 @@ import java.util.List;
 public interface UserMapper {
 
     @Select("select * from online_shopping_user where user_id = #{user_id} and name = #{name}")
-    List<User> selectList(@Param("user_id") Integer user_id, @Param("name") String name);
+    List<User> selectList(@Param("user_id") Long user_id, @Param("name") String name);
 
     @Select("select * from online_shopping_user where user_id = #{user_id}")
-    User selectOne(@Param("user_id") Integer user_id);
+    User selectOne(@Param("user_id") Long user_id);
 
 }
